Move block ID passing from hash to endpoint
This will allow us to reference the same icons regardless of whether we
want to lock the block or not.

In order to do this, we're now using EJS to send data from express to
the app, so we no longer need the appID call anymore either (yay!).
Also, to save on some express routing logic, we're no longer including
styles in webpack, but instead linking directly, since we needed to
handle the svg icon references anyway.